rename package and add comments
