--- pkg/controller/add_cloudwatchmetrics.go
@@ -1,10 +1,0 @@
-package controller
-
-import (
-	"github.com/integr8ly/cloud-resource-operator/pkg/controller/cloudwatchmetrics"
-)
-
-func init() {
-	// AddToManagerFuncs is a list of functions to create controllers and add them to a manager.
-	AddToManagerFuncs = append(AddToManagerFuncs, cloudwatchmetrics.Add)
-}
